Update file names, headers and repo structure
