+++ include/panel.h
@@ -1,5 +1,5 @@
 /*
-** PERSONNAL PROJECT, 2023
+** EPITECH PROJECT, 2023
 ** Libmycsfml
 ** File description:
 ** Header for panels
@@ -47,7 +47,12 @@ panel_t *panel_create(rectransform_t *rect, ptype_t type, void *data);
 void panel_add_childs(panel_t *panel, panel_t **childs, int count);
 
 void panel_destroy(panel_t *panel);
-
+void on_panel_enter(panel_t *panel);
+void on_panel_leave(panel_t *panel);
+void on_panel_pressed(panel_t *panel);
+void on_panel_released(panel_t *panel);
+void on_panel_select(panel_t *panel);
+void on_panel_unselect(panel_t *panel);
 void panel_resize(panel_t *panel, sfVector2f *pos, sfVector2f *size);
 void panel_draw(sfRenderWindow *window, panel_t *panel);
