+++ include/ui_panels/text.h
@@ -1,5 +1,5 @@
 /*
-** PERSONNAL PROJECT, 2023
+** EPITECH PROJECT, 2023
 ** Libmycsfml
 ** File description:
 ** Header for text panels
@@ -7,7 +7,7 @@
 
 #ifndef TEXT_H
     #define TEXT_H
-    #include "panel/panel.h"
+    #include "panel.h"
 
 typedef struct panel_text_s {
     const char *text;
