+++ src/panel/panel_hover.c
@@ -1,8 +1,8 @@
 /*
-** PERSONNAL PROJECT, 2023
+** EPITECH PROJECT, 2023
 ** Libmycsfml
 ** File description:
-** General foos for panels
+** Hovering foos for panels
 */
 
 #include "theme.h"
