+++ src/panel/panel_resize.c
@@ -1,12 +1,12 @@
 /*
-** PERSONNAL PROJECT, 2023
+** EPITECH PROJECT, 2023
 ** Libmycsfml
 ** File description:
-** General foos for panel
+** Resizing foos for panel
 */
 
 #include <stdio.h>
-#include "panel/panel.h"
+#include "panel.h"
 
 static void sprite_rect_set(panel_t *panel)
 {
