+++ src/scene/scene_update.c
@@ -1,8 +1,8 @@
 /*
-** PERSONNAL PROJECT, 2023
+** EPITECH PROJECT, 2023
 ** Libmycsfml
 ** File description:
-** extra functions for the csfml program
+** Update foos for scenes
 */
 
 #include "scene.h"
